refactor: turned html template files used in tab generator page from const-embedding in typescript to .template.html files
There is a considerable drop in performance since the .template.html
files can only be read by server actions or server components, which
sits on the server. Because the html preview is updated live as the user
is typing, an HTTP request is sent to the server every time the tab
title or tab content is changed; hence, the latency.